chore(v3): update starter dependencies
this commit updates the dependencies of the stencil component starter
template in preparation for the the v3.0.0 release of stencil.

the dependency on stencil is bumped to a pre-release version of the
library, and puppeteer is bumped to v19, which is now supported by
stencil v3.

jest dependencies have also been bumped, although that was done during
the course of other package updates (and is not strictly necessary for
this effort).