fix(puppeteer): temporarily pin puppeteer version
with this commit, we pin puppeteer to v21.1.1. we began to
observe that our CI pipeline in Stencil core began to fail
due to type errors like the following:

```
[ ERROR ]  TypeScript: node_modules/puppeteer/lib/types.d.ts:345:15
           Property 'asyncDispose' does not exist on type 'SymbolConstructor'.

    L344:      [Symbol_2.dispose](): void;
    L345:      [Symbol_2.asyncDispose](): Promise<void>;
    L346:  }
```

this occurred immediately following the puppeteer v21.2
release. 

we've confirmed that 21.1.1 still works. while we works out a
minimal repro case, we pin the version here to prevent folks from
not being able to build