chore(scripts): remove --docs from build script
remove the --docs flag from the npm build script. we're removing this as
it's tripped a few people up as to how docs are getting generated when
they remove `docs-readme` from the `stencil-config.ts` file, and still
see documentation generated as a result of this flag.

`docs-readme` already exists in the output targets of
`stencil-config.ts`, which results in no output changes for the
boilerplate project